Refactor/cw721 query (#861)
